+++ src/consts.ts
@@ -149,6 +149,7 @@ export const defaultProjectSettings: Settings = {
   showConnections: "selected",
   showCollisionSlopeTiles: false,
   showCollisionExtraTiles: false,
+  collisionLayerOpacity: 50,
   worldScrollX: 0,
   worldScrollY: 0,
   zoom: 100,